Consolidating some files.
