+++ clients/snowflake/snowflake.go
@@ -2,6 +2,9 @@ package snowflake
 
 import (
 	"context"
+	"fmt"
+
+	"github.com/artie-labs/transfer/clients/utils"
 
 	"github.com/artie-labs/transfer/lib/config"
 	"github.com/artie-labs/transfer/lib/config/constants"
@@ -9,6 +12,7 @@ import (
 	"github.com/artie-labs/transfer/lib/dwh/types"
 	"github.com/artie-labs/transfer/lib/logger"
 	"github.com/artie-labs/transfer/lib/optimization"
+	"github.com/artie-labs/transfer/lib/ptr"
 	"github.com/snowflakedb/gosnowflake"
 )
 
@@ -24,6 +28,20 @@ const (
 	describeTypeCol    = "type"
 	describeCommentCol = "comment"
 )
+
+func (s *Store) getTableConfig(ctx context.Context, fqName string, dropDeletedColumns bool) (*types.DwhTableConfig, error) {
+	return utils.GetTableConfig(ctx, utils.GetTableCfgArgs{
+		Dwh:                s,
+		FqName:             fqName,
+		ConfigMap:          s.configMap,
+		Query:              fmt.Sprintf("DESC table %s;", fqName),
+		ColumnNameLabel:    describeNameCol,
+		ColumnTypeLabel:    describeTypeCol,
+		ColumnDescLabel:    describeCommentCol,
+		EmptyCommentValue:  ptr.ToString("<nil>"),
+		DropDeletedColumns: dropDeletedColumns,
+	})
+}
 
 func (s *Store) Label() constants.DestinationKind {
 	return constants.Snowflake
